@@ -121,9 +121,185 @@ async def extract_tables(file_path: str | PathLike[str], config: GMFTConfig | No
     Returns:
         A list of table data dictionaries.
     """
+    from pathlib import Path
+
+    from kreuzberg._utils._cache import get_table_cache
+
+    # Determine if we should use isolated process  # ~keep
+    if use_isolated_process is None:
+        use_isolated_process = os.environ.get("KREUZBERG_GMFT_ISOLATED", "true").lower() in ("true", "1", "yes")
+
+    path = Path(file_path)
     try:
-        from gmft.auto import AutoTableDetector, AutoTableFormatter
-        from gmft.detectors.tatr import TATRDetectorConfig
+        stat = path.stat()
+        file_info = {
+            "path": str(path.resolve()),
+            "size": stat.st_size,
+            "mtime": stat.st_mtime,
+        }
+    except OSError:
+        file_info = {
+            "path": str(path),
+            "size": 0,
+            "mtime": 0,
+        }
+
+    config = config or GMFTConfig()
+    cache_kwargs = {
+        "file_info": str(sorted(file_info.items())),
+        "extractor": "gmft",
+        "config": str(sorted(config.__dict__.items())),
+    }
+
+    table_cache = get_table_cache()
+    cached_result = await table_cache.aget(**cache_kwargs)
+    if cached_result is not None:
+        return cached_result  # type: ignore[no-any-return]
+
+    if table_cache.is_processing(**cache_kwargs):
+        import anyio
+
+        event = table_cache.mark_processing(**cache_kwargs)
+        await anyio.to_thread.run_sync(event.wait)
+
+        # Try cache again after waiting for other process to complete  # ~keep
+        cached_result = await table_cache.aget(**cache_kwargs)
+        if cached_result is not None:
+            return cached_result  # type: ignore[no-any-return]
+
+    table_cache.mark_processing(**cache_kwargs)
+
+    try:
+        if use_isolated_process:
+            from kreuzberg._multiprocessing.gmft_isolated import extract_tables_isolated_async
+
+            result = await extract_tables_isolated_async(file_path, config)
+
+            await table_cache.aset(result, **cache_kwargs)
+
+            return result
+
+        try:
+            from gmft.auto import AutoTableDetector, AutoTableFormatter  # type: ignore[attr-defined]
+            from gmft.detectors.tatr import TATRDetectorConfig  # type: ignore[attr-defined]
+            from gmft.formatters.tatr import TATRFormatConfig
+            from gmft.pdf_bindings.pdfium import PyPDFium2Document
+
+            formatter: Any = AutoTableFormatter(  # type: ignore[no-untyped-call]
+                config=TATRFormatConfig(
+                    verbosity=config.verbosity,
+                    formatter_base_threshold=config.formatter_base_threshold,
+                    cell_required_confidence=config.cell_required_confidence,
+                    remove_null_rows=config.remove_null_rows,
+                    enable_multi_header=config.enable_multi_header,
+                    semantic_spanning_cells=config.semantic_spanning_cells,
+                    semantic_hierarchical_left_fill=config.semantic_hierarchical_left_fill,
+                    large_table_if_n_rows_removed=config.large_table_if_n_rows_removed,
+                    large_table_threshold=config.large_table_threshold,
+                    large_table_row_overlap_threshold=config.large_table_row_overlap_threshold,
+                    large_table_maximum_rows=config.large_table_maximum_rows,
+                    force_large_table_assumption=config.force_large_table_assumption,
+                )
+            )
+            detector: Any = AutoTableDetector(  # type: ignore[no-untyped-call]
+                config=TATRDetectorConfig(detector_base_threshold=config.detector_base_threshold)
+            )
+            doc = await run_sync(PyPDFium2Document, str(file_path))
+            cropped_tables: list[CroppedTable] = []
+            dataframes: list[DataFrame] = []
+            try:
+                for page in doc:
+                    cropped_tables.extend(await run_sync(detector.extract, page))
+
+                for cropped_table in cropped_tables:
+                    formatted_table = await run_sync(formatter.extract, cropped_table)
+                    dataframes.append(await run_sync(formatted_table.df))
+
+                result = [
+                    TableData(
+                        cropped_image=cropped_table.image(),
+                        page_number=cropped_table.page.page_number,
+                        text=data_frame.to_markdown(),
+                        df=data_frame,
+                    )
+                    for data_frame, cropped_table in zip(dataframes, cropped_tables)
+                ]
+
+                await table_cache.aset(result, **cache_kwargs)
+
+                return result
+            finally:
+                await run_sync(doc.close)
+
+        except ImportError as e:
+            raise MissingDependencyError.create_for_package(
+                dependency_group="gmft", functionality="table extraction", package_name="gmft"
+            ) from e
+    finally:
+        table_cache.mark_complete(**cache_kwargs)
+
+
+def extract_tables_sync(
+    file_path: str | PathLike[str], config: GMFTConfig | None = None, use_isolated_process: bool | None = None
+) -> list[TableData]:
+    """Synchronous wrapper for extract_tables.
+
+    Args:
+        file_path: The path to the PDF file.
+        config: An optional configuration object.
+        use_isolated_process: Whether to use an isolated process for extraction.
+            If None, uses environment variable KREUZBERG_GMFT_ISOLATED (default: True).
+
+    Returns:
+        A list of table data dictionaries.
+    """
+    from pathlib import Path
+
+    from kreuzberg._utils._cache import get_table_cache
+
+    # Determine if we should use isolated process  # ~keep
+    if use_isolated_process is None:
+        use_isolated_process = os.environ.get("KREUZBERG_GMFT_ISOLATED", "true").lower() in ("true", "1", "yes")
+
+    path = Path(file_path)
+    try:
+        stat = path.stat()
+        file_info = {
+            "path": str(path.resolve()),
+            "size": stat.st_size,
+            "mtime": stat.st_mtime,
+        }
+    except OSError:
+        file_info = {
+            "path": str(path),
+            "size": 0,
+            "mtime": 0,
+        }
+
+    config = config or GMFTConfig()
+    cache_kwargs = {
+        "file_info": str(sorted(file_info.items())),
+        "extractor": "gmft",
+        "config": str(sorted(config.__dict__.items())),
+    }
+
+    table_cache = get_table_cache()
+    cached_result = table_cache.get(**cache_kwargs)
+    if cached_result is not None:
+        return cached_result  # type: ignore[no-any-return]
+
+    if use_isolated_process:
+        from kreuzberg._multiprocessing.gmft_isolated import extract_tables_isolated
+
+        result = extract_tables_isolated(file_path, config)
+
+        table_cache.set(result, **cache_kwargs)
+
+        return result
+
+    try:
+        from gmft.auto import AutoTableDetector, AutoTableFormatter  # type: ignore[attr-defined]
+        from gmft.detectors.tatr import TATRDetectorConfig  # type: ignore[attr-defined]
         from gmft.formatters.tatr import TATRFormatConfig
         from gmft.pdf_bindings.pdfium import PyPDFium2Document
 
@@ -145,16 +321,16 @@ async def extract_tables(file_path: str | PathLike[str], config: GMFTConfig | No
             )
         )
         detector = AutoTableDetector(config=TATRDetectorConfig(detector_base_threshold=config.detector_base_threshold))
-        doc = await run_sync(PyPDFium2Document, str(file_path))
+        doc = run_sync(PyPDFium2Document, str(file_path))
         cropped_tables: list[CroppedTable] = []
         dataframes: list[DataFrame] = []
         try:
             for page in doc:
-                cropped_tables.extend(await run_sync(detector.extract, page))
+                cropped_tables.extend(run_sync(detector.extract, page))
 
             for cropped_table in cropped_tables:
-                formatted_table = await run_sync(formatter.extract, cropped_table)
-                dataframes.append(await run_sync(formatted_table.df))
+                formatted_table = run_sync(formatter.extract, cropped_table)
+                dataframes.append(run_sync(formatted_table.df))
 
             return [
                 TableData(
@@ -163,10 +339,10 @@ async def extract_tables(file_path: str | PathLike[str], config: GMFTConfig | No
                     text=data_frame.to_markdown(),
                     df=data_frame,
                 )
-                for data_frame, cropped_table in zip(dataframes, cropped_tables)
+                for data_frame, cropped_table in zip(dataframes, cropped_tables, strict=False)
             ]
         finally:
-            await run_sync(doc.close)
+            run_sync(doc.close)
 
     except ImportError as e:
         raise MissingDependencyError.create_for_package(
